msm7x30: Camera HAL for HTC Sense devices (1/2)
The proprietary camera files must be from an HTC Sense build for this
HAL to function properly. These files include:

libcamera.so
libmmipl.so
libmmjpeg.so
liboemcamera.so

Face unlock and panorama are supported, if available.

Required:
http://review.cyanogenmod.com/#change,13317
http://review.cyanogenmod.com/#change,13285

Optional (preview issues):
http://review.cyanogenmod.com/#change,13564

Notes:
CAF parameters are provided to support building the CAF camera
variation, but will not interfere with regular functionality. They are
simply enclosed for organization